refactor: Move DTOs into respective subpackages
- Created new packages for the DTOs that are specific to their purpose.
  The dto package was starting to be populated with many DTOs and
  it became more difficult to see at a glance which DTOs were related.
  Packaging them solved the problem.

- Due to the reorganization of the dto package, imports had to be changed
  in controllers and services that uses the DTOs.
