+++ backend/src/main/java/dev/jordanjoseph/backend/dto/account/AccountView.java
@@ -1,4 +1,4 @@
-package dev.jordanjoseph.backend.dto;
+package dev.jordanjoseph.backend.dto.account;
 
 import java.math.BigDecimal;
 import java.util.UUID;
