+++ backend/src/main/java/dev/jordanjoseph/backend/dto/authentication/AuthResponse.java
@@ -1,4 +1,4 @@
-package dev.jordanjoseph.backend.dto;
+package dev.jordanjoseph.backend.dto.authentication;
 
 public record AuthResponse(String accessToken, String refreshToken) {
 }
