+++ backend/src/main/java/dev/jordanjoseph/backend/dto/authentication/RefreshRequest.java
@@ -1,4 +1,4 @@
-package dev.jordanjoseph.backend.dto;
+package dev.jordanjoseph.backend.dto.authentication;
 
 public record RefreshRequest(String refreshToken) {
 }
